+++ packages/core/src/__tests__/react-state-mobx.test.ts
@@ -1,0 +1,12 @@
+import { componentToReact } from '../generators/react';
+import { runTestsForTarget } from './test-generator';
+
+describe('React - stateType: mobx', () => {
+  runTestsForTarget({
+    options: {
+      stateType: 'mobx',
+    },
+    target: 'react',
+    generator: componentToReact,
+  });
+});