+++ packages/core/src/__tests__/react.test.ts
@@ -4,8 +4,14 @@ import { runTestsForTarget } from './test-generator';
 
 import stamped from './data/blocks/stamped-io.raw.tsx?raw';
 
-describe('React', () => {
-  runTestsForTarget({ options: {}, target: 'react', generator: componentToReact });
+describe('React - stateType: useState', () => {
+  runTestsForTarget({
+    options: {
+      stateType: 'useState',
+    },
+    target: 'react',
+    generator: componentToReact,
+  });
   test('stamped (useState)', () => {
     const component = parseJsx(stamped);
     const output = componentToReact({