+++ packages/core/src/__tests__/test-generator.ts
@@ -1,8 +1,9 @@
 import { parseJsx } from '@/parsers/jsx';
 import { Target } from '@/types/config';
 import { BaseTranspilerOptions, TranspilerGenerator } from '@/types/transpiler';
-import { describe, test } from 'vitest';
+import { describe, expect, test } from 'vitest';
 import { MitosisComponent, createTypescriptProject, parseSvelte } from '..';
+
 const getRawFile = async (filePath: string) => {
   const code = await import(`${filePath}?raw`).then((x) => x.default as string);
   return { code, filePath: ['src', '__tests__', filePath].join('/') };
@@ -11,6 +12,8 @@ const getRawFile = async (filePath: string) => {
 type RawFile = ReturnType<typeof getRawFile>;
 
 const getterState = getRawFile('./data/blocks/getter-state.raw.tsx');
+const advancedFor = getRawFile('./data/advanced-for.raw.tsx');
+
 const basicForShow = getRawFile('./data/basic-for-show.raw.tsx');
 const basicBooleanAttribute = getRawFile('./data/basic-boolean-attribute.raw.tsx');
 const basicOnMountUpdate = getRawFile('./data/basic-onMount-update.raw.tsx');
@@ -81,6 +84,7 @@ const inputParentBlock = getRawFile('./data/blocks/input-parent.raw.tsx');
 const multipleOnUpdate = getRawFile('./data/blocks/multiple-onUpdate.raw.tsx');
 const multipleOnUpdateWithDeps = getRawFile('./data/blocks/multiple-onUpdateWithDeps.raw.tsx');
 const onInit = getRawFile('./data/blocks/onInit.raw.tsx');
+const onInitPlain = getRawFile('./data/blocks/onInit-plain.raw.tsx');
 const onEvent = getRawFile('./data/blocks/onEvent.raw.tsx');
 const onInitonMount = getRawFile('./data/blocks/onInit-onMount.raw.tsx');
 const onMount = getRawFile('./data/blocks/onMount.raw.tsx');
@@ -131,6 +135,7 @@ const NestedStore = getRawFile('./data/store/nested-store.raw.tsx');
 const UseValueAndFnFromStore = getRawFile('./data/store/use-value-and-fn-from-store.raw.tsx');
 const StringLiteralStore = getRawFile('./data/store/string-literal-store.raw.tsx');
 const StringLiteralStoreKebab = getRawFile('./data/store/string-literal-store-kebab.raw.tsx');
+const StoreAsyncFunction = getRawFile('./data/store-async-function.raw.tsx');
 
 /**
  * Use TestsWithFailFor when you want to write a test that you know will fail
@@ -195,6 +200,7 @@ const BASIC_TESTS: Tests = {
   Columns: columns,
   onUpdate: onUpdate,
   onInit: onInit,
+  onInitPlain,
   onEvent,
   onUpdateWithDeps: onUpdateWithDeps,
   onMount: onMount,
@@ -249,6 +255,7 @@ const BASIC_TESTS: Tests = {
   getterState,
   NestedStore,
   UseValueAndFnFromStore,
+  'store-async-function': StoreAsyncFunction,
   'string-literal-store': StringLiteralStore,
   'string-literal-store-kebab': {
     file: StringLiteralStoreKebab,
@@ -295,6 +302,7 @@ const FORM_BLOCK_TESTS: Tests = {
 const FOR_SHOW_TESTS: Tests = {
   Section: sectionState,
   Basic: basicForShow,
+  Advanced: advancedFor,
 };
 
 const FORWARD_REF_TESTS: Tests = {
@@ -430,6 +438,18 @@ const JSX_TESTS_FOR_TARGET: Partial<Record<Target, Tests[]>> = {
     FOR_SHOW_TESTS,
     ADVANCED_REF,
     ON_UPDATE_RETURN,
+  ],
+  mitosis: [
+    BASIC_TESTS,
+    SLOTS_TESTS,
+    SHOW_TESTS,
+    FORWARD_REF_TESTS,
+    MULTI_ON_UPDATE_TESTS,
+    FORM_BLOCK_TESTS,
+    ADVANCED_REF,
+    ON_UPDATE_RETURN,
+    FOR_SHOW_TESTS,
+    CONTEXT_TEST,
   ],
   webcomponent: [
     CONTEXT_TEST,
@@ -651,7 +671,7 @@ export const runTestsForTarget = <X extends BaseTranspilerOptions>({
       },
       {
         name: 'svelte',
-        parser: async ({ filePath, code }) => parseSvelte(code),
+        parser: async ({ code }) => parseSvelte(code),
         testsArray: filterTests([SVELTE_SYNTAX_TESTS], only),
       },
     ];