+++ packages/core/src/generators/react-native/index.ts
@@ -13,7 +13,7 @@ import { camelCase, size } from 'lodash';
 import traverse from 'neotraverse/legacy';
 import { MitosisNode, Plugin } from '../..';
 import { VALID_HTML_TAGS } from '../../constants/html_tags';
-import { componentToReact } from '../react';
+import { ToReactOptions, componentToReact } from '../react';
 import { sanitizeReactNativeBlockStyles } from './sanitize-react-native-block-styles';
 
 const stylePropertiesThatMustBeNumber = new Set(['lineHeight']);
@@ -30,7 +30,10 @@ const sanitizeStyle = (obj: any) => (key: string, value: string) => {
   }
 };
 
-export const collectReactNativeStyles = (json: MitosisComponent): ClassStyleMap => {
+export const collectReactNativeStyles = (
+  json: MitosisComponent,
+  options: ToReactOptions,
+): ClassStyleMap => {
   const styleMap: ClassStyleMap = {};
 
   const componentIndexes: Dictionary<number | undefined> = {};
@@ -51,7 +54,7 @@ export const collectReactNativeStyles = (json: MitosisComponent): ClassStyleMap 
       // Style properties like `"20px"` need to be numbers like `20` for react native
       for (const key in cssValue) {
         sanitizeStyle(cssValue)(key, cssValue[key]);
-        cssValue = sanitizeReactNativeBlockStyles(cssValue);
+        cssValue = sanitizeReactNativeBlockStyles(cssValue, options);
       }
     }
 
@@ -61,7 +64,7 @@ export const collectReactNativeStyles = (json: MitosisComponent): ClassStyleMap 
         // Style properties like `"20px"` need to be numbers like `20` for react native
         for (const key in styleValue) {
           sanitizeStyle(styleValue)(key, styleValue[key]);
-          styleValue = sanitizeReactNativeBlockStyles(styleValue);
+          styleValue = sanitizeReactNativeBlockStyles(styleValue, options);
         }
 
         item.bindings.style!.code = json5.stringify(styleValue);