+++ packages/core/src/generators/react-native/types.ts
@@ -1,6 +1,7 @@
 import { BaseTranspilerOptions } from '@/types/transpiler';
 
 export interface ToReactNativeOptions extends BaseTranspilerOptions {
+  sanitizeReactNative?: boolean;
   stylesType: 'emotion' | 'react-native' | 'twrnc' | 'native-wind';
   stateType: 'useState' | 'mobx' | 'valtio' | 'solid' | 'builder';
 }