+++ packages/core/src/generators/react/generator.ts
@@ -242,7 +242,9 @@ export const componentToReact: TranspilerGenerator<Partial<ToReactOptions>> =
           // Remove spaces between imports
           .replace(/;\n\nimport\s/g, ';\nimport ');
       } catch (err) {
-        console.error('Format error for file:', str);
+        if (process.env.NODE_ENV !== 'test') {
+          console.error('Format error for file:', str);
+        }
         throw err;
       }
     }
@@ -289,18 +291,21 @@ const getPropsDefinition = ({ json }: { json: MitosisComponent }) => {
   return `${json.name}.defaultProps = {${defaultPropsString}};`;
 };
 
+const isRSC = (json: MitosisComponent, options: ToReactOptions) => {
+  // When using RSC generator, we check `componentType` field in metadata to determine if it's a server component
+  const componentType = json.meta.useMetadata?.rsc?.componentType;
+  if (options.rsc && checkIsDefined(componentType)) {
+    return componentType === 'server';
+  }
+
+  return !checkIfIsClientComponent(json);
+};
 const checkShouldAddUseClientDirective = (json: MitosisComponent, options: ToReactOptions) => {
   if (!options.addUseClientDirectiveIfNeeded) return false;
   if (options.type === 'native') return false;
   if (options.preact) return false;
 
-  // When using RSC generator, we check `componentType` field in metadata to determine if it's a server component
-  const componentType = json.meta.useMetadata?.rsc?.componentType;
-  if (options.rsc && checkIsDefined(componentType)) {
-    return componentType === 'client';
-  }
-
-  return checkIfIsClientComponent(json);
+  return !isRSC(json, options);
 };
 
 const _componentToReact = (
@@ -376,7 +381,13 @@ const _componentToReact = (
   if (hasContext(json) && options.contextType !== 'prop-drill') {
     reactLibImports.add('useContext');
   }
-  if (allRefs.length || json.hooks.onInit?.code) {
+
+  const shouldAddUseClientDirective = checkShouldAddUseClientDirective(json, options);
+
+  const shouldInlineOnInitHook =
+    !shouldAddUseClientDirective && options.rsc && isRSC(json, options);
+
+  if (allRefs.length || (json.hooks.onInit?.code && !shouldInlineOnInitHook)) {
     reactLibImports.add('useRef');
   }
   if (!options.preact && hasPropRef) {
@@ -405,7 +416,7 @@ const _componentToReact = (
   // side effects that delete styles bindings from the JSON.
   const reactNativeStyles =
     options.stylesType === 'react-native' && componentHasStyles
-      ? collectReactNativeStyles(json)
+      ? collectReactNativeStyles(json, options)
       : undefined;
 
   const propType = json.propsTypeRef || 'any';
@@ -436,7 +447,8 @@ const _componentToReact = (
               keyPrefix: 'const',
               valueMapper: (code, type, _, key) => {
                 if (type === 'getter') return `${key} = function ${code.replace('get ', '')}`;
-                if (type === 'function') return `${key} = function ${code}`;
+                if (type === 'function')
+                  return code.startsWith('async') ? code : `${key} = function ${code}`;
                 return code;
               },
             })
@@ -450,7 +462,9 @@ const _componentToReact = (
 
     ${
       json.hooks.onInit?.code
-        ? `
+        ? shouldInlineOnInitHook
+          ? processHookCode({ str: json.hooks.onInit.code, options })
+          : `
         const hasInitialized = useRef(false);
         if (!hasInitialized.current) {
           ${processHookCode({
@@ -529,8 +543,6 @@ const _componentToReact = (
       ${wrap ? closeFrag(options) : ''}
     );
   `;
-
-  const shouldAddUseClientDirective = checkShouldAddUseClientDirective(json, options);
 
   const str = dedent`
   ${shouldAddUseClientDirective ? `'use client';` : ''}