+++ packages/core/src/helpers/bindings.ts
@@ -1,6 +1,11 @@
 import { Binding } from '../types/mitosis-node';
 
-export const createSingleBinding = (args: Omit<Binding, 'type'>): Binding => ({
+type SingleBinding = Omit<Binding & { type: 'single' }, 'type'>;
+
+export const createSingleBinding = (
+  args: Omit<SingleBinding, 'bindingType'> & Partial<Pick<SingleBinding, 'bindingType'>>,
+): Binding => ({
   ...args,
+  bindingType: args.bindingType || 'expression',
   type: 'single',
 });