+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,110 @@
 # @builder.io/mitosis-cli
 
+## 0.5.15
+
+### Patch Changes
+
+- Updated dependencies [a0ad5ab]
+  - @builder.io/mitosis@0.5.15
+
+## 0.5.14
+
+### Patch Changes
+
+- Updated dependencies [39af4d6]
+  - @builder.io/mitosis@0.5.14
+
+## 0.5.13
+
+### Patch Changes
+
+- Updated dependencies [c7d2f8c]
+  - @builder.io/mitosis@0.5.13
+
+## 0.5.12
+
+### Patch Changes
+
+- Updated dependencies [5e2cf3c]
+  - @builder.io/mitosis@0.5.12
+
+## 0.5.11
+
+### Patch Changes
+
+- Updated dependencies [db9dbf9]
+  - @builder.io/mitosis@0.5.11
+
+## 0.5.10
+
+### Patch Changes
+
+- Updated dependencies [499b4b7]
+  - @builder.io/mitosis@0.5.10
+
+## 0.5.9
+
+### Patch Changes
+
+- Updated dependencies [8c2be87]
+  - @builder.io/mitosis@0.5.9
+
+## 0.5.8
+
+### Patch Changes
+
+- Updated dependencies [8d823a1]
+  - @builder.io/mitosis@0.5.8
+
+## 0.5.7
+
+### Patch Changes
+
+- Updated dependencies [7a099d2]
+  - @builder.io/mitosis@0.5.7
+
+## 0.5.6
+
+### Patch Changes
+
+- Updated dependencies [0aa642a]
+  - @builder.io/mitosis@0.5.6
+
+## 0.5.5
+
+### Patch Changes
+
+- Updated dependencies [56c6347]
+  - @builder.io/mitosis@0.5.5
+
+## 0.5.4
+
+### Patch Changes
+
+- Updated dependencies [d13e693]
+  - @builder.io/mitosis@0.5.4
+
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies [82c79fd]
+  - @builder.io/mitosis@0.5.3
+
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [dad8e2f]
+  - @builder.io/mitosis@0.5.2
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [dbd5a50]
+  - @builder.io/mitosis@0.5.1
+
 ## 0.5.0
 
 ### Patch Changes